state: describe backend with phandle
We used to describe the backend with a combination of a nodepath
and partname:<name>. This is cumbersome and error prone to parse.
Instead, describe the backend with a phandle that directly points
to the partition. The resulting code is easier to read and the
devicetrees are easier to understand. Tested with both mtd and
EEPROM device, both with partitions with an offset != 0.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>