barebox-state: make state_get() definition consistent to its prototype
In 2839d4f6f4 state_get() got an additional argument, but its prototype
in barebox-state.h was never adjusted. Include the file containing
the prototype from the file containing the function definition so that
this bug can be detected by the compiler. Also add the missing argument
to the callers.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>