barebox-state: block while waiting for file lock
In multi-processes environment, it can happen that barebox-state is
called concurrently. Since barebox-state usually only runs for at most a
few seconds, it is probably okay for the second process to block while
wait for the file lock, so the calling entity (user or script) doesn't
have to do error handling itself.

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>