common: align declarations of dev_add_param_*() functions with barebox
The only remaining difference is the argument "value" in dev_add_param_enum()
which is of type int* in barebox and got type uint32_t* here to not let the
compiler warn about assignment of pointers of different signedness.

There is no such warning when compiling barebox as it's applying the compiler
option "-Wno-pointer-sign" in its main makefile.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Signed-off-by: Roland Hieber <rhi@pengutronix.de>