libdt: drop broken if-branch
device_find_block_device returns 0 on success, so the way the else if
clause is now, only if there is a block device, the code falls through.
If there is none, a 0 is returned, but devpath is not populated breaking
the contract of the function. Just drop the branch for now and add it back
later in a way that works.

Fixes: 929ed64cb42f ("of_get_devicepath: make partition finding more robust")
Reviewed-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Link: https://lore.pengutronix.de/oss-tools/20230607121628.3180139-4-a.fatoum@pengutronix.de
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>