libdt: use block device partition instead of parent if found
Linux doesn't parse MTD fixed-partitions described in the device tree
for block devices and EEPROMs. For block devices, The dt-utils work around
this by doing the lookup in userspace and then arrive at the parent block
device along with an offset that is passed along. We can do better
though: If there's a partition block device that contains the region
we are interested in, we should just be using that.

This avoids the issue of triggering a reread of the partition table
after writing barebox state because udev is notified that we had
been opening the whole block device writable.

Tested-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Link: https://lore.pengutronix.de/oss-tools/20230607121628.3180139-6-a.fatoum@pengutronix.de
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>