libdt: use of_find_device_by_uuid for partuuid lookup
With the addition of of_find_device_by_uuid as part of the support for
barebox,storage-by-uuid, we don't need to rely on the device symlinks
created by 60-persistent-storage.rules and instead can just use libudev
directly, which we do elsewhere anyway. This has the added benefit
that we unify with the other code paths, where we determine the device
path through the struct udev_device.

Link: https://lore.pengutronix.de/oss-tools/20230607121628.3180139-8-a.fatoum@pengutronix.de
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>