treewide: add SPDX identifiers to files with GPL-2.0-only license
Move the GPL-2.0 license text to the new LICENSES folder, where it will
reside next to all other licenses. Add the rest of the license too (see
FSF FAQ [1]) and clean up the hard page breaks in the process.

Add the standard license dedication and warranty disclaimer too for
completeness sake, so we can easily reference the full license terms by
using its SPDX identifier, and don't have to carry around the
boilerplate headers in the future. However, keep the already existing
license headers in files, as recommended in the REUSE FAQ [2].

"COPYING" is probably still referenced somewhere, so make it point to
the new location of the GPL.

Link: [1] https://www.gnu.org/licenses/gpl-faq.html#GPLOmitPreamble
Link: [2] https://reuse.software/faq/#edit-copyright-and-licensing
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/oss-tools/20230731091131.3696307-2-rhi@pengutronix.de
Signed-off-by: Roland Hieber <rhi@pengutronix.de>