Makefile: include meson files in the dist tarball
Currently the dist tarball still needs to be built by autotools because
it adds all the generated files in build-aux/ and m4/ needed for
configure, and the pre-built configure script and Makefile as well.
However add all the files needed for meson build as well so the tarball
can be used with both build systems.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>