README: provide git format.subjectPrefix line to copy
The oss-tools list is shared between a number of projects, so we should
always use a subject prefix to make it clear where the patch applies.

The README already notes that and references the git man page, but it's
more convenient to just provide a command line that can be directly
copied and pasted.

Cc: Bastian Krause <bst@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/oss-tools/20231106112953.3994614-1-a.fatoum@pengutronix.de
Signed-off-by: Roland Hieber <rhi@pengutronix.de>